Refactor job adapters metric collection logging to a shared method
All logging from the job adapters follow the same pattern, so move that
logic to a shared method that can create the log message by the end of
the metrics collection the same way across them, which simplifies the
bulk logic of the collection methods.

This will also enable us to do further refactorings in the future, for
example avoiding to build the log messages if the log level won't allow
it.