Fix tests to run without redis online
It turns out my first attempt to implement these tests worked because I
had an instance of redis running, and so some internal attempts from
Resque to store and fetch data worked, causing the tests to pass. Thanks
for the build for letting me know pretty quickly that was incorrect.

In order to bypass Redis altogether, this creates a "stub"
implementation of Resque::Worker for our tests, similar to how we stub
Sidekiq queues.

I think we should consider running against real Redis eventually, like
we need to run DJ or Que against a real DB, but I'll leave that for a
later time.