Setup adapter configs as instances that receive the identifier
Instead of passing around the identifier + config class and having the
config be responsible for instantiating those, instantiate them upfront
when setting up the adapter, make it receive the identifier, so the
config carries its identifier with it to make it easier to access, and
just set things up with the config instance directly.

This required us to create a similar `reset` method for that config,
because we need a clean state between tests, but it makes it more inline
with how the config as a whole work.