Refactor config json moving the identifier to the job adapter config
This makes the job adapter config responsible for emitting the JSON with
the identifier as key, instead of making that responsibility of the
external config object. This way, each config handles their own
attributes internally and the whole representation, which makes the
config more similar to the adapters when dumping for sending back to the
API.

Also refactor both JSON generation pieces to simplify with `reduce`,
since they use similar implementation now.