Add latency suppor to Resque via an extension / plugin / monkey-patch
Queue latency is a much better metric to perform scaling on, but Resque
doesn't have a built-in way of calculating that, only the depth / size
of a queue is provided.

This extends Resque to track a new "timestamp" attribute with each job
enqueued / pushed to the queue, and use that timestamp information from
the oldest job present in the queue when we ask for the latency of that
queue, which is essentially the time difference between now and that
oldest job timestamp.

This will allow Judoscale to better handle scaling for Resque workers,
as well as make Resque work like the other libraries that already
provide latency and scale based on that.

I have not removed queue depth metric tracking here, but this is
something that may happen in the future.