fix: Don't start the reporter in build processes (#180)
* Don't start the reporter in build processes

During build processes such as asset precompiling or database migrations, there's no need to report metrics. We've seen instances of these reports causing unexpected scaling.

There's no surefire way to know when we're in a build process, so we're checking for the presence of a Rake task. We need to allow for rake tasks that run jobs, though, and we do that through the `config.allow_rake_tasks` option. This is set automatically for DelayedJob and Resque.

We're also no longer including the middleware when running a Rails console. This wasn't causing problems, but it added noise to the console output.

* Fix tests

* Require "set" for older Ruby versions