Copy over full GoodJob migration to silence warnings
GoodJob was complaining that some migrations needed to be executed,
apparently because we were only creating the main `good_jobs` table and
associated indexes.

That table has now a few more columns and there's a few other tables, to
keep it simple I copied over the full thing, which makes it easier to
update again in the future as well.

I also updated the DelayedJob migration with the index creation that was
missing, just to keep it a full raw copy of the migration similar to
GoodJob.