Test against Que v1.x
The main Gemfile on judoscale-que will install the latest Que version,
which means it should be covering v2.x, so the secondary que-2 Gemfile
was likely testing a similar version.

We need a separate Gemfile that enforces installation of Que v1.x in
order to test it against judoscale-que. v1.x doesn't support Ruby 3+
though, so we only test it against Ruby 2.x and the related Active
Record versions that support the same Ruby.