chore: Test matrix & sample app dependency updates & cleanup (#197)
* Bundle update all sample apps, make sure they are all running/working

* Expand Ruby support via test coverage

* Test against Rails / DH / GJ / Que against Active Record v7.0

The main Gemfile points to AR latest, which is currently v7.1, and we
have tests against v6.1, this adds testing for the intermediate v7.0 as
well to make sure we cover all those versions.

* Run tests for against Sidekiq v6

The main Gemfile is testing Sidekiq v7, we have a separate Gemfile for
v5, this adds v6 as well.

* Update to the latest checkout action

This fixes action warnings in GH about old node versions used by the
checkout@v2. The latest v4 is using node20 and should eliminate all
those warnings.

    The following actions uses node12 which is deprecated and will be forced to run on node16: actions/checkout@v2.
    For more info: https://github.blog/changelog/2023-06-13-github-actions-all-actions-will-run-on-node16-instead-of-node12-by-default/

    Node.js 16 actions are deprecated. Please update the following actions to use Node.js 20: actions/checkout@v2.
    For more information see: https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/.

* Copy over full GoodJob migration to silence warnings

GoodJob was complaining that some migrations needed to be executed,
apparently because we were only creating the main `good_jobs` table and
associated indexes.

That table has now a few more columns and there's a few other tables, to
keep it simple I copied over the full thing, which makes it easier to
update again in the future as well.

I also updated the DelayedJob migration with the index creation that was
missing, just to keep it a full raw copy of the migration similar to
GoodJob.

* Supress migration messages from DJ/GJ

Make the test output less noisy without those migration messages,
specially now that GJ is larger with all the tables/indexes.

* Skip Ruby 3.2+ and AR 6.1 tests on Good Job

We're getting an error due to Ruby 3.2+ and kwargs apparently, some
Rails 6.1 incompatibility with latest Ruby versions.

    ArgumentError: wrong number of arguments (given 1, expected 0)
        test/metrics_collector_test.rb:57:in `block (3 levels) in <module:Judoscale>'

Rails has fixed most of those on 6-1-stable branch, but no stable release
is planned any time soon (that branch is receiving security fixes only),
so let's just skip instead of running everything against the branch for
simplicity, since we're running our tests on other versions and it's a
known Rails 6-1 issue.

* Test against Que v1.x

The main Gemfile on judoscale-que will install the latest Que version,
which means it should be covering v2.x, so the secondary que-2 Gemfile
was likely testing a similar version.

We need a separate Gemfile that enforces installation of Que v1.x in
order to test it against judoscale-que. v1.x doesn't support Ruby 3+
though, so we only test it against Ruby 2.x and the related Active
Record versions that support the same Ruby.

* Switch database names on sample apps / test dbs back to judoscale_*

* Update some Rails Autoscale instances to Judoscale

Some comments / tests / sample apps copy were referencing Rails
Autoscale from a previous renaming, update those to Judoscale to match
the code / libs across the board.

Rails Autoscale is supported alongside Judoscale, but we're referencing
things as Judoscale.

* Switch sample apps to use request catcher

The previous request inspector endpoint seems to have expired and we'd
have to generate a new one, and request catcher seems to be working as
before, so switching it back for simplicity since we can just hit
whatever subdomain there and it shows up.

* Document the new/correct port that changed with heroku local

* Remove invalid references for application stylesheets

These apps have no styles, so no such file exists. This was turning into
a 404 request shown in the logs.