Use Shoryuken cache for fetching queues
Every queue is initialized with a name, and it will fetch the queue URL
from SQS upon init, causing us to reach out to SQS twice on every
collection: one to translate queue name => url, and another to fetch the
metadata we need to report queue depth.

By using the cached queues from Shoryuken, we don't reach out to SQS as
much, making the collection process faster as well.