Implement direct call to SQS to fetch the queue depth attribute value
Shoryuken provides a private method/API to fetch queue attributes, in
which we were relying before, but reaching out to private APIs doesn't
feel really great.

Instead, we're reaching out to SQS directly using a similar
implementation of the private `Shoryuken::Queue#queue_attributes`, but
fetching only the attribute we need for reporting.