Simplify logic to gather queues
We don't need an intermediate hash to initialize the queue names, since
we're working off `Shoryuken.ungrouped_queues` always, and we rely on
Shoryuken cached queues, we can just fetch the queues during the loop,
and remove the extraneous initialization.

There's also no need to initialize the queue names with the metrics
collector, since it's the same call used during collection, the first
known queues will be initialized there.