Setup steps after release-please to automate gem version syncing
This is an attempt to automatically sync our gem versions after
release-please detects the need to bump the version. It will open up a
PR with that new version bump to judoscale-ruby and changelog added, and
these extra steps should follow-up that with syncing the new version
across the other libraries.

I have done some manual testing on this using a separate branch, by
commenting out the release-please and "if" statements, and only
triggering these actual steps that sync versions and push a new commit.
By bumping the judoscale-ruby version myself in a commit, the automated
steps here would generate a new follow-up commit automatically syncing
the other gems versions, so I'm hopeful this will work as expected.
(famous last words)

The git user name/email and commands in general are based on examples:
https://github.com/actions/checkout?tab=readme-ov-file#push-a-commit-using-the-built-in-token
https://github.com/google-github-actions/release-please-action/tree/v3.7.13?tab=readme-ov-file#creating-majorminor-tags