Checkout the PR head branch name where we need to sync
We need to checkout the release PR to sync versions, otherwise it might
use `main` which is where this workflow is running.

It seems the "outputs" object has access to a PR object if one was
created (which it'd have been in this case):
https://github.com/google-github-actions/release-please-action?tab=readme-ov-file#outputs

and that PR object has access to both head & base branch names:
https://github.com/googleapis/release-please/blob/cb106cd0b5ae5c3e2a3b015487dc615eba72b842/src/pull-request.ts#L15-L17

so I believe we can checkout the right branch for that PR and run the
sync/commit there, with this change.

It's hard to test all of this without actually running a release though,
so we might have to wait and see! (and hope!)