Make rails-autoscale-* require judoscale-* base lib file
Instead of directly requiring judoscale/* files, make it require the
base judoscale-* lib file, and leave the work of requiring judoscale/*
files to judoscale-* itself.

This is more flexible as it allows judoscale-* to have extra code/setup
if necessary (as it happens with judoscale-ruby for instance, which
rails-autoscale-core is currently requiring), and might allow us to make
rails-autoscale-* gems to be more like shims that depend on judoscale-*
in the future.