Make rails-autoscale-* depend sole on their judoscale-* counterpart
The rails-autoscale-* gems will be shims on top of judoscale-* gems,
simply depending on / requiring them going forward.

We only include the rails-autoscale-* files on the gem via gemspec now,
and the specific `version.rb` file for each, since that's necessary for
the dependency / gemspec setup.

This means the judoscale-* gems can now dictate the dependencies without
us having to update both judoscale-* and rails-autoscale-* versions.