Simplify gem versioning
Rely on a single VERSION constant from judoscale-ruby, require that
relatively to the specific lib gemspec so we can properly build the gem
when releasing (which gets generated automatically via release-please).

This means we don't need to "sync" gem versions anymore manually, or try
to automate that as part of our release-please workflows, since there's
a single source of truth for the version, and all libs share / release
the same version anyway, even if there are no changes, let's just always
require / use that single version file.