fix: lock GoodJob dependency to < 4.0
GoodJob v4 was recently released, but that has some breaking changes
that don't work well with Judoscale yet. Before adding full support to
the new version, we're doing a patch release to lock it to version < 4,
so that people trying to install both don't run into problems using
GoodJob v4 and Judoscale.

https://island94.org/2024/07/introducing-goodjob-v4