Mark the reporter thread as "fork safe" to avoid warnings
This prevents the Reporter thread from generating warnings with puma,
which has been warning on any threads created before forking.

The Reporter thread is aware of forking by keeping the current `pid`
around, and initializing another thread if the `pid` changes. It means
we still have a running Reporter on the main puma process in this case,
plus one for each forked process (if running in cluster mode, in case of
puma.)

References:
https://github.com/rails/rails/pull/40399
https://github.com/puma/puma/pull/2475
https://github.com/puma/puma/releases/tag/v5.1.0

Closes #170