Load entire SolidQueue schema file manually into our test DB
SolidQueue v0.8+ merged all the migrations into a single schema file
that is automatically loaded onto the DB on a Rails app. They suggest a
separate DB, which AR will handle automatically by copying over this
single file schema to the app, and to use on the same DB one would have
to manually move the schema to a migration and execute it.

For our case, all we care about is to have SolidQueue schema / tables
loaded onto our test DB. We do that by manually loading the lib's schema
file via Active Record. We still execute migrations afterwards, even
though it's a no-op for current SolidQueue, if they need to make new
schema changes in the future, they'll come in the form of migrations.