Stop appending the legacy Render service ID to the API url
We support it on the backend for legacy installations that migrate over
to `JUDOSCALE_URL` so they can keep using the same gems, but there's no
need to continue using the legacy URLs on new versions now.

If the app is setup with `JUDOSCALE_URL` it should just use that, if
it's setup with the legacy render support nothing changes, they report
the Render service ID as their token.