Move reporter start to the beginning of the middleware flow
It's a little odd that it sits in between two bits of code that are
about tracking, so it should be either at the beginning, or the end. I
chose the beginning, for Rails apps this is just an "ensure running"
call because the reporter should really be started from the initializer
and from non-Rails apps this might actually start the reporter on the
first hit, but keeping it at the top doesn't change the current behavior
since it still happens before we push metrics to the store.