Cache started at time object
Started at timestamp doesn't change, but it's called twice when
calculating queue time (on an initial check to make sure it's available,
then on the body of the method).

We can cache it to avoid doing extra checks and creating an extra time
object on each request we track queue time.

It can also be moved to private, since it's not really used by the code
or tests.