Work with UTC values to calculate queue time
Consistently use UTC to calculate queue time, regardless of time zone
the app's servers are set to.

Previously we were using non-UTC times, as long as they were the same
unit it was fine... now we're using UTC times, again it's the same unit
so we're fine. But since we can pass this as an argument (and we do pass
it now from our middleware), it's better to ensure they are the same
unit and the best option is to keep all UTC.