refactor: Report the same time across all metrics collected on each adapter (#237)
We had some metrics being reported with a shared "time" value that's
initialized at the beginning of the collection methods, while other
metrics had their own instance of time during the loop that fetches and
pushes/queues the metrics.

This makes it a bit more consistent by always initializing the time at
the beginning of the collection, and sharing that same time value for
all created metrics within that collection loop for each adapter.