Always track app time, regardless of queue time availability
We want to always track & report app time, regardless of having queue
time available (or skipping it due to large payloads), so that it can
help with debugging in some cases. They don't need to be linked to each
other for reporting, i.e. no need to skip app time because we are not
tracking queue time.