feat: Track app time as a new metric (#238)
This can be useful to help determine internally if queue time increase
has direct relation to app time increase, or actual capacity issues, so
we'll track it as a new metric. 

We want to always track & report app time, regardless of having queue
time available (or skipping it due to large payloads), so that it can
help with debugging in some cases as well.