Introduce new middleware to track & collect utilization metrics
The middleware will track a counter of inflight requests and push active
requests and whether that represents an active process with a configured
interval. We'll explore that information to be used as an "utilization"
autoscaling metric.

Those metrics go to the same store as the other metrics (queue time,
etc.) so that they're included with the reports to our backend
automatically.

Note that the middleware is only available in judoscale-rails for now,
and is not included by default, it needs to be manually enabled via
config or ENV var.