Create a "foreach" bin script that to run commands on all adapter gems
This makes it easier to run simple commands like bundle install/update
across all adapter gems.

Also separate the update script for dependencies of the gem and
dependencies of the sample app. Often times I just want to get the
adapters updated and end up causing a lot of diffs on the sample apps
with dependencies I don't care at the moment, so splitting will allow
more control over what to update and when.