Lock minitest to the latest version that supports Ruby 2.6
They've released 5.25.5 dropping support to Ruby 2.6, but we still
support it in our test suite, so lock to the previous version to be able
to run those tests on a few of our adapters.

Some adapters don't run the main Gemfile against Ruby 2.6 because of
other dependencies like Rails/AR that only work with newer versions, but
for those that do, we need this version lock. The change is made across
the board to test with the same version.