Stop running utilization tracking tests through the thread
There seems to be some weird thing going on with this test when running
via the utilization tracker thread and the loop/sleep stubs on Ruby 2.7,
it randomly fails as if `Thread#join` didn't do anything / waited for
the thread to run/finish. This test fails on CI under Ruby 2.7 on almost
every run, changing between AR 7.0 and AR 7.1 versions, but always on
Ruby 2.7, and it never failed on other versions of Ruby.

Instead of running through the thread, let's assume the thread is doing
the right thing, and just test that the tracking logic works / pushes
metrics to be reported as we expect.