Update Gemfile.lock files during release (#244)
This is meant to update the Gemfile.lock files to the new version that's
being created, otherwise they'll become invalid when we merge the
release PR, causing test failures and requiring us to manually run bundle
install, as it happened with v1.11.0.

I tried a few combinations using release-please "updater" concept via
the `GemfileLock` that comes with Ruby, but it only seems to apply if
having a single `Gemfile.lock` file in the main repo, or if we used
their multirepo setup. Since none of those worked I went with a manual
attempt by setting up Ruby/bundle and running install & commit.