Attempt to fix release-please check for PR created
I mistakenly looked up the v4 of the release-please-action which
contains the `prs_created` output flag
https://github.com/googleapis/release-please-action?tab=readme-ov-file#outputs

However, we're using v3 which does not apparently:
https://github.com/googleapis/release-please-action/tree/v3.7.13?tab=readme-ov-file#outputs

This switches to check for the presence of the `pr` JSON string object,
which identifies the created PR if any, and should be undefined if no
release was created. We won't be able to test this until a new version
gets released anyway, we'll see how it goes.