feat: Track utilization more accurately (#248)
This replaces the "process utilization" (pu) with "utilization percentage" (up). Conceptually they are the same, but the implementation is completely different.

Utilization percentage tracks an "idle_ratio", which is essentially the inverse of utilization. We calculate it as we go since we always know when the web process is idle. This avoids the need for sampling, eliminating the extra thread loop and providing a more precise metric for utilization.