Eliminate the need to convert back to the logger severity identifier
We were doing back-and-forth conversion here with the logger level:
first we'd coerce it to an integer if necessary, than back to a symbol,
or just working with integers all along, but the back-and-forth seems
unnecessary if we can just "add" to the logger by passing the level /
severity as an argument, rather than using `public_send` for the given
level.

All loggers should implement `add`/`log` the same, hopefully.